Use the default SSLContext for DefaultSslContextFactory
Use the JVM's SSLContext default instead of manually interpreting the defaults and specifically calling out a "TLS" (which is TLSv1) provider.

For some version of IBM Java (6 & 7 tested), the TLSv1.2 provider did not also contain a "TLSv1" implementation. Setting system properties to enforce TLSv1.2 on IBM Java created an error because the SSLContext initialized here could not use TLSv1.2.